project built, cannot populate trending and album click

+++ app/src/main/java/com/example/melody_meter_local/model/PopularSearch.kt
@@ -2,5 +2,5 @@ package com.example.melody_meter_local.model
 
 data class PopularSearch(
     val searchString: String = "",
-    var count: Int = 0
+    var count: Long = 0
 )
